streamlit_cloud/requirements.txt
streamlit
chromadb
pypdf
python-docx
tiktoken
requests
pandas
tenacity
pysqlite3-binary
